refactor: split reader/reader.go — read operations into reader_read.go
